Added MAXALLSHIPS! option to ship list
Added a MAXALLSHIPS! option to the ship list.

This change touched a couple of elements to reduce duplication of code. Notably the creation of  both a FindShips and MaxLevel function.

The ship list is now a list which is printed with a seperator, this list is then used by MAXALLSHIPS! to iterate through each ship.